+++ src/exceptions/abstract.exception.ts
@@ -1,0 +1,17 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+import { ContextString } from '../common/types';
+
+export class AbstractException extends HttpException {
+
+    constructor(
+        private readonly context: ContextString,
+        message: string,
+        private readonly statusCode: HttpStatus
+    ) {
+        super(message, statusCode);
+    }
+
+    public getContext(): ContextString {
+        return this.context;
+    }
+}