+++ src/main.ts
@@ -1,12 +1,15 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import * as cookieParser from 'cookie-parser';
-import { ValidationPipe } from '@nestjs/common';
+import { HttpExceptionFilter } from './filters/http-exception.filter';
+import { ValidationPipe } from './pipes/validation.pipe';
 
 async function bootstrap() {
     const app = await NestFactory.create(AppModule);
     app.use(cookieParser());
+    app.useGlobalFilters(new HttpExceptionFilter())
     app.useGlobalPipes(new ValidationPipe());
     await app.listen(3000);
 }
+
 bootstrap();