+++ src/modules/auth/auth.controller.ts
@@ -1,6 +1,6 @@
 import { Controller } from '@nestjs/common';
-import { AuthResult } from './auth.interface';
 import { AuthService } from './auth.service';
+import { UserDocument } from '../user/user.interface';
 
 @Controller('auth')
 export class AuthController {
@@ -8,7 +8,7 @@ export class AuthController {
     constructor(private readonly authService: AuthService) {
     }
 
-    async decode(jwtCookie: string): Promise<AuthResult> {
-        return await this.authService.getAnalysis(jwtCookie);
+    async decode(jwtCookie: string): Promise<UserDocument> {
+        return await this.authService.getAuthorizedUser(jwtCookie);
     }
 }