+++ src/pipes/validation.pipe.ts
@@ -1,0 +1,33 @@
+import { PipeTransform, Injectable, ArgumentMetadata } from '@nestjs/common';
+import { validate } from 'class-validator';
+import { plainToInstance } from 'class-transformer';
+import { BadRequestException } from '../exceptions/bad-request.exception';
+
+@Injectable()
+export class ValidationPipe implements PipeTransform<any> {
+
+    async transform(value: any, { metatype }: ArgumentMetadata) {
+        if (!metatype || !this.toValidate(metatype)) {
+            return value;
+        }
+
+        const object = plainToInstance(metatype, value);
+        const errors = await validate(object);
+
+        if (errors.length > 0) {
+            const context = 'ValidationPipe/transform';
+            const message = errors.join('');
+            console.error(context, message);
+
+            throw new BadRequestException(context, message);
+        }
+
+        return value;
+    }
+
+    private toValidate(metatype: Function): boolean {
+        const types: Function[] = [String, Boolean, Number, Array, Object];
+        
+        return !types.includes(metatype);
+    }
+}