+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2023-09-01
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): HTTP exceptions.
+- [Marek Kurańda](https://github.com/mjkuranda): Custom validation pipe.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): Own implementation of status codes.
+
 ## [1.8.1] - 2023-08-31
 ### Fixed
 - [Marek Kurańda](https://github.com/mjkuranda): Payload definitions in API section in readme file.